Struct servo_arc::Arc

#[repr(C)]pub struct Arc<T: ?Sized + 'static> { /* fields omitted */ }

Implementations

impl<T> Arc<T>

pub fn new(data: T) -> Self

pub fn into_raw(this: Self) -> *const T

pub fn borrow_arc<'a>(&'a self) -> ArcBorrow<'a, T>

Produce a pointer to the data that can be converted back to an arc

pub fn with_raw_offset_arc<F, U>(&self, f: F) -> U where
F: FnOnce(&RawOffsetArc<T>) -> U,

Temporarily converts |self| into a bonafide RawOffsetArc and exposes it to the provided callback. The refcount is not modified.

pub fn heap_ptr(&self) -> *const c_void

Returns the address on the heap of the Arc itself – not the T within it – for memory reporting.

impl<T: ?Sized> Arc<T>

pub fn ptr_eq(this: &Self, other: &Self) -> bool

impl<T: Clone> Arc<T>

pub fn make_mut(this: &mut Self) -> &mut T

impl<T: ?Sized> Arc<T>

pub fn get_mut(this: &mut Self) -> Option<&mut T>

pub fn is_unique(&self) -> bool

impl<H, T> Arc<HeaderSlice<H, [T]>>

pub fn from_header_and_iter<I>(header: H, items: I) -> Self where
I: Iterator<Item = T> + ExactSizeIterator,

Creates an Arc for a HeaderSlice using the given header struct and iterator to generate the slice. The resulting Arc will be fat.

impl<H: 'static, T: 'static> Arc<HeaderSlice<HeaderWithLength<H>, [T]>>

pub fn into_thin(a: Self) -> ThinArc<H, T>

Converts an Arc into a ThinArc. This consumes the Arc, so the refcount is not modified.

pub fn from_thin(a: ThinArc<H, T>) -> Self

Converts a ThinArc into an Arc. This consumes the ThinArc, so the refcount is not modified.

impl<T: 'static> Arc<T>

pub fn into_raw_offset(a: Self) -> RawOffsetArc<T>

Converts an Arc into a RawOffsetArc. This consumes the Arc, so the refcount is not modified.

pub fn from_raw_offset(a: RawOffsetArc<T>) -> Self

Converts a RawOffsetArc into an Arc. This consumes the RawOffsetArc, so the refcount is not modified.