Struct servo_arc::ArcBorrow

pub struct ArcBorrow<'a, T: 'a>(_);

A "borrowed Arc". This is a pointer to a T that is known to have been allocated within an Arc.

This is equivalent in guarantees to &Arc<T>, however it is a bit more flexible. To obtain an &Arc<T> you must have an Arc instance somewhere pinned down until we're done with it.

However, Gecko hands us refcounted things as pointers to T directly, so we have to conjure up a temporary Arc on the stack each time. The same happens for when the object is managed by a RawOffsetArc.

ArcBorrow lets us deal with borrows of known-refcounted objects without needing to worry about how they're actually stored.

Methods

impl<'a, T> ArcBorrow<'a, T>

pub fn clone_arc(&self) -> Arc<T>

pub unsafe fn from_ref(r: &'a T) -> Self

For constructing from a reference known to be Arc-backed, e.g. if we obtain such a reference over FFI

pub fn with_arc<F, U>(&self, f: F) -> U where
F: FnOnce(&Arc<T>) -> U,
T: 'static,