servo_arc

Struct ArcUnion

pub struct ArcUnion<A, B> {
    pub(crate) p: NonNull<()>,
    pub(crate) phantom_a: PhantomData<A>,
    pub(crate) phantom_b: PhantomData<B>,
}

A tagged union that can represent Arc<A> or Arc<B> while only consuming a single word. The type is also NonNull, and thus can be stored in an Option without increasing size.

This is functionally equivalent to enum ArcUnion<A, B> { First(Arc<A>), Second(Arc<B>) but only takes up up a single word of stack space.

This could probably be extended to support four types if necessary.

Fields
p: NonNull<()>phantom_a: PhantomData<A>phantom_b: PhantomData<B>
Implementations

impl<A, B> ArcUnion<A, B>

pub(crate) unsafe fn new(ptr: *mut ()) -> Self

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two values are pointer-equal.

pub fn ptr(&self) -> NonNull<()>

pub fn borrow(&self) -> ArcUnionBorrow<'_, A, B>

Returns an enum representing a borrow of either A or B.

pub fn from_first(other: Arc<A>) -> Self

Creates an ArcUnion from an instance of the first type.

pub fn from_second(other: Arc<B>) -> Self

Creates an ArcUnion from an instance of the second type.

pub fn is_first(&self) -> bool

Returns true if this ArcUnion contains the first type.

pub fn is_second(&self) -> bool

Returns true if this ArcUnion contains the second type.

pub fn as_first(&self) -> Option<ArcBorrow<'_, A>>

Returns a borrow of the first type if applicable, otherwise None.

pub fn as_second(&self) -> Option<ArcBorrow<'_, B>>

Returns a borrow of the second type if applicable, otherwise None.