Struct servo_arc::RawOffsetArc

#[repr(C)]

pub struct RawOffsetArc<T: 'static> { /* fields omitted */ }

An Arc, except it holds a pointer to the T instead of to the entire ArcInner.

 Arc<T>    RawOffsetArc<T>
  |          |
  v          v
 ---------------------
| RefCount | T (data) | [ArcInner<T>]
 ---------------------

This means that this is a direct pointer to its contained data (and can be read from by both C++ and Rust), but we can also convert it to a "regular" Arc by removing the offset

Methods

impl<T: 'static> RawOffsetArc<T>

pub fn with_arc<F, U>(&self, f: F) -> U where
F: FnOnce(&Arc<T>) -> U,

Temporarily converts |self| into a bonafide Arc and exposes it to the provided callback. The refcount is not modified.

pub fn make_mut(&mut self) -> &mut T where
T: Clone,

If uniquely owned, provide a mutable reference Else create a copy, and mutate that

pub fn clone_arc(&self) -> Arc<T>

Clone it as an Arc

pub fn borrow_arc<'a>(&'a self) -> ArcBorrow<'a, T>

Produce a pointer to the data that can be converted back to an arc