Struct servo_arc::ThinArc

pub struct ThinArc<H: 'static, T: 'static> { /* fields omitted */ }

Methods

impl<H: 'static, T: 'static> ThinArc<H, T>

pub fn with_arc<F, U>(&self, f: F) -> U where
F: FnOnce(&Arc<HeaderSlice<HeaderWithLength<H>, [T]>>) -> U,

Temporarily converts |self| into a bonafide Arc and exposes it to the provided callback. The refcount is not modified.

pub fn heap_ptr(&self) -> *const c_void

Returns the address on the heap of the ThinArc itself -- not the T within it -- for memory reporting.