Struct servo_arc::Arc

#[repr(C)]
pub struct Arc<T: ?Sized> {
    pub(crate) p: NonNull<ArcInner<T>>,
    pub(crate) phantom: PhantomData<T>,
}

An atomically reference counted shared pointer

See the documentation for Arc in the standard library. Unlike the standard library Arc, this Arc does not support weak reference counting.

See the discussion in https://github.com/rust-lang/rust/pull/60594 for the usage of PhantomData.

cbindgen:derive-eq=false cbindgen:derive-neq=false

Fields
p: NonNull<ArcInner<T>>phantom: PhantomData<T>
Implementations

impl<T> Arc<T>

pub fn new(data: T) -> Self

Construct an Arc<T>

pub fn new_leaked(data: T) -> Self

Construct an intentionally-leaked arc.

pub(crate) fn into_raw(this: Self) -> *const T

Convert the Arc to a raw pointer, suitable for use across FFI

Note: This returns a pointer to the data T, which is offset in the allocation.

It is recommended to use RawOffsetArc for this.

pub(crate) unsafe fn from_raw(ptr: *const T) -> Self

Reconstruct the Arc from a raw pointer obtained from into_raw()

Note: This raw pointer will be offset in the allocation and must be preceded by the atomic count.

It is recommended to use RawOffsetArc for this

pub unsafe fn new_static<F>(alloc: F, data: T) -> Arc<T> where
F: FnOnce(Layout) -> *mut u8,

Create a new static Arc (one that won’t reference count the object) and place it in the allocation provided by the specified alloc function.

alloc must return a pointer into a static allocation suitable for storing data with the Layout passed into it. The pointer returned by alloc will not be freed.

pub fn borrow_arc<'a>(&'a self) -> ArcBorrow<'a, T>

Produce a pointer to the data that can be converted back to an Arc. This is basically an &Arc<T>, without the extra indirection. It has the benefits of an &T but also knows about the underlying refcount and can be converted into more Arc<T>s if necessary.

pub fn with_raw_offset_arc<F, U>(&self, f: F) -> U where
F: FnOnce(&RawOffsetArc<T>) -> U,

Temporarily converts |self| into a bonafide RawOffsetArc and exposes it to the provided callback. The refcount is not modified.

pub fn heap_ptr(&self) -> *const c_void

Returns the address on the heap of the Arc itself – not the T within it – for memory reporting.

If this is a static reference, this returns null.

impl<T: ?Sized> Arc<T>

pub(crate) fn inner(&self) -> &ArcInner<T>

pub(crate) fn record_drop(&self)

pub fn mark_as_intentionally_leaked(&self)

Marks this Arc as intentionally leaked for the purposes of refcount logging.

It’s a logic error to call this more than once, but it’s not unsafe, as it’d just report negative leaks.

pub(crate) unsafe fn drop_slow(&mut self)

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Test pointer equality between the two Arcs, i.e. they must be the same allocation

pub(crate) fn ptr(&self) -> *mut ArcInner<T>

impl<T: Clone> Arc<T>

pub fn make_mut(this: &mut Self) -> &mut T

Makes a mutable reference to the Arc, cloning if necessary

This is functionally equivalent to Arc::make_mut from the standard library.

If this Arc is uniquely owned, make_mut() will provide a mutable reference to the contents. If not, make_mut() will create a new Arc with a copy of the contents, update this to point to it, and provide a mutable reference to its contents.

This is useful for implementing copy-on-write schemes where you wish to avoid copying things if your Arc is not shared.

impl<T: ?Sized> Arc<T>

pub fn get_mut(this: &mut Self) -> Option<&mut T>

Provides mutable access to the contents if the Arc is uniquely owned.

pub fn is_static(&self) -> bool

Whether or not the Arc is a static reference.

pub fn is_unique(&self) -> bool

Whether or not the Arc is uniquely owned (is the refcount 1?) and not a static reference.

impl<H, T> Arc<HeaderSlice<H, [T]>>

pub(crate) fn from_header_and_iter_alloc<F, I>(
alloc: F,
header: H,
items: I,
is_static: bool
) -> Self where
F: FnOnce(Layout) -> *mut u8,
I: Iterator<Item = T> + ExactSizeIterator,

Creates an Arc for a HeaderSlice using the given header struct and iterator to generate the slice.

is_static indicates whether to create a static Arc.

alloc is used to get a pointer to the memory into which the dynamically sized ArcInner<HeaderSlice<H, T>> value will be written. If is_static is true, then alloc must return a pointer into some static memory allocation. If it is false, then alloc must return an allocation that can be dellocated by calling Box::from_raw::<ArcInner<HeaderSlice<H, T>>> on it.

pub fn from_header_and_iter<I>(header: H, items: I) -> Self where
I: Iterator<Item = T> + ExactSizeIterator,

Creates an Arc for a HeaderSlice using the given header struct and iterator to generate the slice. The resulting Arc will be fat.

pub(crate) unsafe fn allocate_buffer<W>(size: usize) -> *mut u8

impl<H, T> Arc<HeaderSlice<HeaderWithLength<H>, [T]>>

pub fn into_thin(a: Self) -> ThinArc<H, T>

Converts an Arc into a ThinArc. This consumes the Arc, so the refcount is not modified.

pub fn from_thin(a: ThinArc<H, T>) -> Self

Converts a ThinArc into an Arc. This consumes the ThinArc, so the refcount is not modified.

impl<T> Arc<T>

pub fn into_raw_offset(a: Self) -> RawOffsetArc<T>

Converts an Arc into a RawOffsetArc. This consumes the Arc, so the refcount is not modified.

pub fn from_raw_offset(a: RawOffsetArc<T>) -> Self

Converts a RawOffsetArc into an Arc. This consumes the RawOffsetArc, so the refcount is not modified.