Struct servo_arc::UniqueArc

pub struct UniqueArc<T: ?Sized + 'static>(_);

An Arc that is known to be uniquely owned

This lets us build arcs that we can mutate before freezing, without needing to change the allocation

Methods

impl<T> UniqueArc<T>

pub fn new(data: T) -> Self

Construct a new UniqueArc

pub fn shareable(self) -> Arc<T>

Convert to a shareable Arc once we're done using it